python-services/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
dnspython==2.4.2
python-whois==0.9.4
phonenumbers==8.13.27
Pillow==10.2.0
python-multipart==0.0.6
cryptography==42.0.2
pyOpenSSL==24.0.0
aiohttp==3.9.1
python-dotenv==1.0.0
slowapi==0.1.9
email-validator==2.1.0
